# github_runner/api/openapi.yaml
openapi: 3.0.3
info:
  title: GitHub Runner Manager API
  version: 0.6.0
  description: |
    Control-plane API for Docker-based GitHub Actions self-hosted runners.

    Runner containers default to the community image myoung34/github-runner
    (https://github.com/myoung34/docker-github-actions-runner) — this API manages
    those containers; it does not implement the Actions runner agent.

    Success responses use `{ "result": "ok", "data": … }`.
    Errors use `{ "result": "error", "error": { "code", "message", "details" } }`.

    Authentication is network-trust (Home Assistant ingress or private bind).
    Do not expose this API on a public interface.

    Optional `GITHUB_PAT` mints registration tokens and deregisters runners on delete.
    Live logs are also available over WebSocket at `/ws` (channel `container_logs`).

    WebSocket client messages:
    `{ "type": "subscribe", "channel": "container_logs", "runner_id": "<id>", "tail": "200" }`
    `{ "type": "unsubscribe", "channel": "container_logs" }`
servers:
  - url: /api/v1
paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthEnvelope"
  /runners:
    get:
      summary: List runners
      operationId: listRunners
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerListEnvelope"
    post:
      summary: Create runner
      operationId: createRunner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunnerRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerEnvelope"
        "400":
          description: Validation error (including missing :local image)
        "409":
          description: Duplicate name or container
        "429":
          description: Rate limited
        "502":
          description: GitHub API error or image pull failure
        "503":
          description: Docker unavailable
  /runners/recreate-missing:
    post:
      summary: Recreate every runner whose container is missing (operator fleet restore)
      operationId: recreateMissingRunners
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecreateRunnerRequest"
      responses:
        "200":
          description: OK (partial failures are listed in data.failed; reconcile stays read-only)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecreateMissingEnvelope"
        "400":
          description: Validation error
        "429":
          description: Rate limited (one quota token for the batch)
        "502":
          description: GitHub API error or image pull failure
        "503":
          description: Docker unavailable
  /runners/{id}:
    get:
      summary: Get runner
      operationId: getRunner
      parameters:
        - $ref: "#/components/parameters/RunnerId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerEnvelope"
        "404":
          description: Not found
    patch:
      summary: Update runner config (labels / runtime). Set apply=true to recreate container.
      operationId: patchRunner
      parameters:
        - $ref: "#/components/parameters/RunnerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchRunnerRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerEnvelope"
        "400":
          description: Validation error (including missing image or failed configure)
        "404":
          description: Not found
        "409":
          description: Runner busy when apply=true and the managed container is running
        "502":
          description: GitHub API error or image pull failure
        "503":
          description: Docker unavailable
    delete:
      summary: Delete runner (local; GitHub deregister when PAT configured)
      operationId: deleteRunner
      parameters:
        - $ref: "#/components/parameters/RunnerId"
      responses:
        "200":
          description: OK
        "404":
          description: Not found
        "409":
          description: Runner busy (only when the managed container is running)
        "503":
          description: Docker unavailable
  /runners/{id}/start:
    post:
      summary: Start runner container
      operationId: startRunner
      parameters:
        - $ref: "#/components/parameters/RunnerId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerEnvelope"
        "400":
          description: Validation error (container missing — recreate it)
        "404":
          description: Not found
        "503":
          description: Docker unavailable
  /runners/{id}/stop:
    post:
      summary: Stop runner container
      operationId: stopRunner
      parameters:
        - $ref: "#/components/parameters/RunnerId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerEnvelope"
        "404":
          description: Not found
        "503":
          description: Docker unavailable
  /runners/{id}/restart:
    post:
      summary: Restart runner container
      operationId: restartRunner
      parameters:
        - $ref: "#/components/parameters/RunnerId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerEnvelope"
        "400":
          description: Validation error (container missing — recreate it)
        "404":
          description: Not found
        "503":
          description: Docker unavailable
  /runners/{id}/recreate:
    post:
      summary: Recreate runner container (keeps reusable volume credentials; token/PAT when volume missing or empty)
      operationId: recreateRunner
      parameters:
        - $ref: "#/components/parameters/RunnerId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecreateRunnerRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerEnvelope"
        "400":
          description: Validation error (including missing image or failed configure)
        "404":
          description: Not found
        "409":
          description: Runner busy (only when the managed container is running)
        "429":
          description: Rate limited
        "502":
          description: GitHub API error or image pull failure
        "503":
          description: Docker unavailable
  /runners/{id}/logs:
    get:
      summary: Runner container logs (plain-text chunked stream)
      operationId: runnerLogs
      parameters:
        - $ref: "#/components/parameters/RunnerId"
        - name: follow
          in: query
          schema:
            type: boolean
        - name: tail
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: Plain text log stream (demultiplexed stdout/stderr)
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Validation error (container missing — recreate it)
        "404":
          description: Not found
        "503":
          description: Docker unavailable
components:
  parameters:
    RunnerId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    CreateRunnerRequest:
      type: object
      required: [name, url]
      properties:
        name:
          type: string
        url:
          type: string
          description: Org URL (one path segment) or repo URL (owner/repo)
        token:
          type: string
          description: Short-lived registration token; optional when GITHUB_PAT is configured
        labels:
          type: array
          items:
            type: string
        image:
          type: string
          description: Per-runner image override (tag or digest)
        cpu_limit:
          type: number
          description: CPU limit in cores (0 = unlimited)
        memory_limit_mb:
          type: integer
          description: Memory limit in MiB (0 = unlimited)
        extra_env:
          type: object
          additionalProperties:
            type: string
          description: Extra container env; values must not contain newline, CR, or NUL
        network_mode:
          type: string
          description: Empty, bridge, host, none, or container:<id>
        mount_docker_sock:
          type: boolean
          description: Override global MOUNT_DOCKER_SOCK for this runner
        cache:
          $ref: "#/components/schemas/CacheConfig"
        workdir_host_path:
          type: string
          description: |
            Absolute Docker-host path same-path bind-mounted as RUNNER_WORKDIR.
            Empty defaults to /srv/gha-work/<normalized-name>. Must not be under
            /var/lib/docker. Changing workdir requires agent reconfigure (token/PAT).
    PatchRunnerRequest:
      type: object
      properties:
        labels:
          type: array
          items:
            type: string
        image:
          type: string
        cpu_limit:
          type: number
        memory_limit_mb:
          type: integer
        extra_env:
          type: object
          additionalProperties:
            type: string
          description: Extra container env; values must not contain newline, CR, or NUL
        network_mode:
          type: string
          description: Empty, bridge, host, none, or container:<id>
        mount_docker_sock:
          type: boolean
        reset_mount_docker_sock:
          type: boolean
          description: Clear per-runner docker-sock override (use global default)
        cache:
          $ref: "#/components/schemas/CacheConfig"
        workdir_host_path:
          type: string
          description: Set empty string to reset to the default /srv/gha-work/<name>
        apply:
          type: boolean
          description: When true, recreate the container to apply runtime changes
        token:
          type: string
          description: |
            Required when apply=true if the data volume is missing or workdir reconfigure
            is needed (agent workFolder ≠ host bind); optional otherwise when PAT is set
    RecreateRunnerRequest:
      type: object
      properties:
        token:
          type: string
          description: |
            Required when the data volume is missing or empty (no .runner), or when
            workFolder does not match the host workdir bind (reconfigure). Optional
            override when GITHUB_PAT is configured. Unused when volume credentials
            can be reused.
    RecreateMissingEnvelope:
      type: object
      properties:
        result:
          type: string
          enum: [ok]
        data:
          type: object
          properties:
            recreated:
              type: array
              items:
                type: string
            failed:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  error:
                    type: string

    RunnerStatus:
      type: string
      enum: [running, exited, missing, unknown]
    JobState:
      type: string
      enum: [idle, busy, unknown]
      description: |
        Actions agent activity when the container is running, from local job hooks
        (status.json). Omitted when the container is not running.
    CurrentJob:
      type: object
      description: Present when job_state is busy (from ACTIONS_RUNNER_HOOK_JOB_* env).
      properties:
        repository:
          type: string
        workflow:
          type: string
        job:
          type: string
        run_id:
          type: string
        run_number:
          type: string
        run_attempt:
          type: string
        sha:
          type: string
        ref:
          type: string
        actor:
          type: string
        event:
          type: string
        updated_at:
          type: string
          format: date-time
    Runner:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
        scope:
          type: string
          enum: [repo, org]
        labels:
          type: array
          items:
            type: string
        container_name:
          type: string
        volume_name:
          type: string
        image:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        cpu_limit:
          type: number
        memory_limit_mb:
          type: integer
        extra_env:
          type: object
          additionalProperties:
            type: string
          description: Extra container env; values must not contain newline, CR, or NUL
        network_mode:
          type: string
          description: Empty, bridge, host, none, or container:<id>
        mount_docker_sock:
          type: boolean
        cache:
          $ref: "#/components/schemas/CacheConfig"
        workdir_host_path:
          type: string
          description: Stored override; empty means default /srv/gha-work/<name>
        workdir_effective:
          type: string
          description: Resolved host path used for RUNNER_WORKDIR and the same-path bind
        workdir_agent:
          type: string
          description: workFolder from /runner/data/.runner (set only at configure time)
        workdir_mismatch:
          type: boolean
          description: True when workdir_agent ≠ workdir_effective
        workdir_error:
          type: string
          description: Present when reading .runner failed
        cache_effective:
          type: string
          description: Resolved cache path injected as RUNNER_CACHE (bind host_path or volume target)
        warnings:
          type: array
          items:
            type: string
          description: Soft configuration advisories (request still succeeds)
        status:
          $ref: "#/components/schemas/RunnerStatus"
        running:
          type: boolean
        job_state:
          $ref: "#/components/schemas/JobState"
        current_job:
          $ref: "#/components/schemas/CurrentJob"
    CacheConfig:
      type: object
      description: |
        Optional persistent cache mount. Use the same volume_name (or host_path)
        on multiple runners to share a cache. UI defaults to host bind; API omit-type
        default remains volume for compatibility. Bind mounts use Docker host paths
        (not paths inside the addon container) and are always same-path (target
        coerced to host_path). The runner container receives RUNNER_CACHE with the
        effective path for workflows and Buildx type=local.

        Named volumes are allowed but surface soft warnings[] — sibling Docker host
        binds cannot see named volumes.
      properties:
        enabled:
          type: boolean
        type:
          type: string
          enum: [volume, bind]
          description: API default volume when enabled; manager UI defaults to bind
        volume_name:
          type: string
          description: Named volume; empty defaults to gha-runner-<name>-cache
        host_path:
          type: string
          description: Absolute Docker host path when type=bind (same-path mount; required)
        target:
          type: string
          description: Container mount path for volume mode (default /cache); ignored/coerced for bind
        read_only:
          type: boolean
    RunnerEnvelope:
      type: object
      properties:
        result:
          type: string
          enum: [ok]
        data:
          $ref: "#/components/schemas/Runner"
    RunnerListEnvelope:
      type: object
      properties:
        result:
          type: string
          enum: [ok]
        data:
          type: object
          properties:
            runners:
              type: array
              items:
                $ref: "#/components/schemas/Runner"
    HealthEnvelope:
      type: object
      properties:
        result:
          type: string
          enum: [ok]
        data:
          type: object
          properties:
            status:
              type: string
              enum: [ok, degraded]
              description: ok when docker and store are healthy; degraded otherwise (HTTP stays 200)
            version:
              type: string
            github_pat_configured:
              type: boolean
            runner_image:
              type: string
            mount_docker_sock:
              type: boolean
            runners:
              type: object
              additionalProperties:
                type: integer
            orphans:
              type: array
              items:
                type: object
            docker:
              type: object
              properties:
                available:
                  type: boolean
                engine:
                  type: string
                error:
                  type: string
            store:
              type: object
              properties:
                readable:
                  type: boolean
                error:
                  type: string